Rename python_object_detector.py
